Add support for ROM images
For the miamil2, everything remains in RAM. The BSP builds a binary that
can be placed into RAM directly.
For the ml506l2, executable code and initialized data are to be placed in
ROM. The BSP builds the ROM contents, the program will copy the data
into RAM at boottime in start.S.
Ran "bootstrap -p" to regenerate preinstall.am files, otherwise the new
files don't actually get installed.