ml506l2: Support running from ROM
Initialize the LEON2 config registers at boot, in particular the MCFG
registers to make RAM work (and speed up ROM) and set the stack pointer
to the end of RAM.

Setup the trap table at the start of the data segment and postpone
setting up trap support until after initializing data and bss.

Had to copy and adjust start.S for that, there appears to be no hook to
put this part in.

Remove the memory-split coherent config since that isn't available yet
in logic.

Tested using the "hello" example application in the "sis" simulator.